Fix XCode 6.3 Build Warnings

+++ XMLRPCConnectionManager.h
@@ -17,7 +17,7 @@
 
 - (NSArray *)activeConnectionIdentifiers;
 
-- (int)numberOfActiveConnections;
+- (NSUInteger)numberOfActiveConnections;
 
 #pragma mark -
